backend/requirements.txt
tweepy
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
psycopg2
email_validator
jinja2
orjson